rename Validation Class to ValidationBase, rename ValidationAlgorithms to Validations
