+++ framework/Models/PostProcessors/Factory.py
@@ -35,8 +35,8 @@ from .ComparisonStatisticsModule import ComparisonStatistics
 from .RealizationAverager import RealizationAverager
 from .ParetoFrontierPostProcessor import ParetoFrontier
 from .EconomicRatio import EconomicRatio
-from .Validation import Validation
-from .ValidationAlgorithms import Probabilistic
+from .ValidationBase import ValidationBase
+from .Validations import Probabilistic
 from .TSACharacterizer import TSACharacterizer
 
 ### PostProcessorFunctions (orig: InterfacedPostProcessor)
